[MU-94] Fix loadAllUser
